+++ amplitude.js
@@ -562,7 +562,9 @@ AmplitudeClient.prototype.init = function init(apiKey, opt_userId, opt_config, o
 
     // load deviceId and userId from input, or try to fetch existing value from cookie
     this.options.deviceId = (type(opt_config) === 'object' && type(opt_config.deviceId) === 'string' &&
-        !utils.isEmptyString(opt_config.deviceId) && opt_config.deviceId) || this.options.deviceId || UUID() + 'R';
+        !utils.isEmptyString(opt_config.deviceId) && opt_config.deviceId) ||
+        (this.options.deviceIdFromUrlParam && this._getDeviceIdFromUrlParam(this._getUrlParams())) ||
+        this.options.deviceId || UUID() + 'R';
     this.options.userId = (type(opt_userId) === 'string' && !utils.isEmptyString(opt_userId) && opt_userId) ||
         this.options.userId || null;
 
@@ -984,6 +986,14 @@ AmplitudeClient.prototype._saveGclid = function _saveGclid(urlParams) {
   }
   var gclidProperties = {'gclid': gclid};
   _sendParamsReferrerUserProperties(this, gclidProperties);
+};
+
+/**
+ * Try to fetch Amplitude device id from url params.
+ * @private
+ */
+AmplitudeClient.prototype._getDeviceIdFromUrlParam = function _getDeviceIdFromUrlParam(urlParams) {
+  return utils.getQueryParam(Constants.AMP_DEVICE_ID_PARAM, urlParams);
 };
 
 /**
@@ -1686,7 +1696,9 @@ module.exports = {
   REVENUE_PRODUCT_ID: '$productId',
   REVENUE_QUANTITY: '$quantity',
   REVENUE_PRICE: '$price',
-  REVENUE_REVENUE_TYPE: '$revenueType'
+  REVENUE_REVENUE_TYPE: '$revenueType',
+
+  AMP_DEVICE_ID_PARAM: 'amp_device_id'  // url param
 };
 
 }, {}],
@@ -4991,7 +5003,8 @@ module.exports = {
   eventUploadPeriodMillis: 30 * 1000, // 30s
   forceHttps: false,
   includeGclid: false,
-  saveParamsReferrerOncePerSession: true
+  saveParamsReferrerOncePerSession: true,
+  deviceIdFromUrlParam: false,
 };
 
 }, {"./language":29}],