+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Add `logEventWithTimestamp` to allow logging events with a custom timestamp. The timestamp should a number representing the time in milliseconds since epoch. See [documentation](https://rawgit.com/amplitude/Amplitude-Javascript/master/documentation/AmplitudeClient.html) for more details.
+* Add configuration option `deviceIdFromUrlParam`, which when set to `true` will have the SDK parse device Ids from url parameter `amp_device_id` if available. DeviceIds defined in the configuration options during init will take priority over device Ids from url parameters.
 
 ### 3.3.2 (October 28, 2016)